requirements.txt
absl-py==2.1.0
asttokens
backcall
blosc2==2.0.0
cachetools==5.5.0
certifi==2024.8.30
chardet==4.0.0
cloudpickle==3.0.0
comm
contourpy==1.1.1
cycler==0.12.1
debugpy
decorator
etils==1.3.0
executing
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
glfw==2.7.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
grpcio-tools==1.47.0
h5py==2.9.0
idna==2.10
importlib_metadata
importlib_resources==6.4.5
ipykernel
ipython
ipywidgets==8.1.5
jedi
Jinja2==3.1.4
joblib==1.4.2
jupyter_client
jupyter_core
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline
mpmath==1.3.0
msgpack==1.1.0
nest_asyncio
networkx==3.1
numexpr==2.8.6
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging
pandas==1.3.5
parso
pexpect
pickleshare
pillow==10.4.0
platformdirs
prompt_toolkit
protobuf==3.19.5
psutil
ptyprocess
pure_eval
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyglet==1.5.26
Pygments
PyOpenGL==3.1.7
pyparsing==3.1.4
python-dateutil
pytz==2024.2
PyYAML==6.0.2
pyzmq
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.10.1
seaborn==0.13.2
six
sk-video==1.1.10
stack-data
sympy==1.13.3
tables==3.8.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
torch==2.4.1
tornado
tqdm==4.67.1
traitlets
triton==3.0.0
typing_extensions
urllib3==1.26.20
wcwidth
Werkzeug==3.0.4
widgetsnbextension==4.0.13
zipp
